requirements.txt
django == 3.2.1
pillow
django-admin-interface
django-crispy-forms
django-rest-framework
social-auth-app-django
django-pwa
django-bootstrap4
django-bootstrap-datepicker-plus
django-jet-reboot
python-decouple